Finalise use of package repo
This is a much cleaner process now. The tailscale package is installed
from the repo and the .deb stored in /config to be reinstalled on
firmware upgrade.

The firstboot and post-config scripts seem to work with minimal (no?)
manual intervention needed and should ensure that Tailscale starts after
routing is available. We now use a bind mount of the state directory to
/var/lib/tailscale as the symlink wasn't reliable - the bind mount means
systemd can ensure it's in place before starting the service.

The post-config script should detect when tailscaled.service needs
restarting and do so, in case unit overrides or the mount point weren't
in place when the package was installed.